Fixed image search for imzml (now relays on identity % to file name)
