+++ mzml2isa/versionutils.py
@@ -7,9 +7,9 @@ the other modules.
 
 About
 -----------------------------------------------------------------------------
-The mzml2isa parser was created by Tom Lawson (University of Birmingham, UK) 
+The mzml2isa parser was created by Tom Lawson (University of Birmingham, UK)
 as part of a NERC funded placement at EBI Cambridge in June 2015. Python 3
-port and small enhancements were carried out by Martin Larralde (ENS Cachan, 
+port and small enhancements were carried out by Martin Larralde (ENS Cachan,
 France) in June 2016 during an internship at the EBI Cambridge.
 
 License
@@ -25,26 +25,26 @@ from copy import deepcopy
 
 if sys.version_info == 2:
     try:
-        from lxml import etree 
+        from lxml import etree
     except ImportError:
-        from xml.etree import cElementTree as etree 
-    
-    def pyxpath(mzMLmeta, query): 
+        from xml.etree import cElementTree as etree
+
+    def pyxpath(mzMLmeta, query):
         """Finds every occurence of *query* in *mzMLmeta.tree* with proper namespace
 
-        This function also formats the xpath query using the *mzMLmeta.env 
+        This function also formats the xpath query using the *mzMLmeta.env
         dictionnary created by the **mzml.mzMLmeta.build_env** function.
-        """ 
+        """
         return mzMLmeta.tree.xpath(query.format(**mzMLmeta.env), namespaces=mzMLmeta.ns)
-    
-    def getparent(element, tree): 
+
+    def getparent(element, tree):
         """Finds every parent of a tree node.
 
         Uses the method provided by lxml.etree
         """
         return element.getparent()
-    
-    def iterdict(dictionary): 
+
+    def iterdict(dictionary):
         """Creates an iterator on the items of a dictionnary"""
         return dictionary.iteritems()
 
@@ -78,25 +78,25 @@ else:
     try:
         from lxml import etree
     except ImportError:
-        from xml.etree import ElementTree as etree 
-    
+        from xml.etree import ElementTree as etree
+
     def pyxpath(mzMLmeta, query):
         """Finds every occurence of *query* in *mzMLmeta.tree* with proper namespace
 
-        This function also formats the xpath query using the *mzMLmeta.env 
+        This function also formats the xpath query using the *mzMLmeta.env
         dictionnary created by the **mzml.mzMLmeta.build_env** function.
-        """ 
+        """
         return mzMLmeta.tree.findall(query.format(**mzMLmeta.env), mzMLmeta.ns)
-    
-    def getparent(element, tree): 
+
+    def getparent(element, tree):
         """Finds every parent of a tree node.
 
         As xml.ElementTree has no **.getparent** method, the following was
         proposed here : http://stackoverflow.com/questions/2170610#20132342
         """
         return {c:p for p in tree.iter() for c in p}[element]
-    
-    def iterdict(dictionary): 
+
+    def iterdict(dictionary):
         """Creates an iterator on the items of a dictionnary"""
         return dictionary.items()
 
@@ -123,3 +123,18 @@ else:
 
     RMODE = 'r'
     WMODE = 'w'
+
+
+
+def longest_substring(string1, string2):
+    answer = ""
+    len1, len2 = len(string1), len(string2)
+    for i in range(len1):
+        match = ""
+        for j in range(len2):
+            if (i + j < len1 and string1[i + j] == string2[j]):
+                match += string2[j]
+            else:
+                if (len(match) > len(answer)): answer = match
+                match = ""
+    return answer
